Fix bug with updating pending function
I wasn't able to figure out how to build the repo (no build scripts in
the repository) so I replaced all files with the content from the npm
package.

The fix is to add a waitFor call in between the updateCode and
updateConfig call on line 158 in ./serverless.js.
Without this change the updateLambda call will fail with the following
error:
```
ResourceConflictException: The operation cannot be performed at this
time. An update is in progress for resource:
arn:aws:lambda:us-east-1:xxxx:function:xxx-xxx
```

See this link for details about the aws change that necessitates this
cahnge:
https://aws.amazon.com/blogs/compute/coming-soon-expansion-of-aws-lambda-states-to-all-functions/